feat: show offers' pins on map
