+++ project/src/components/geo-map/geo-map.tsx
@@ -1,7 +1,7 @@
-import { useEffect, useRef, useState, } from 'react';
-import { Map as LeafletGeoMap, TileLayer } from 'leaflet';
+import { useRef } from 'react';
 import { City, Locations } from 'src/types/types';
-import { LatLngTuple } from 'leaflet';
+import { useGeoMap } from './use-geo-map';
+import { useGeoMapPins } from './use-geo-map-pins';
 import 'leaflet/dist/leaflet.css';
 import styles from './geo-map.module.css';
 
@@ -12,37 +12,9 @@ type GeoMapProps = {
 }
 
 function GeoMap(props: GeoMapProps): JSX.Element {
-  const [ geoMap, setGeoMap ] = useState<LeafletGeoMap | null>(null);
-  const isRenderedRef = useRef<boolean>(false);
   const nodeRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    if (isRenderedRef.current || !nodeRef.current) {
-      return;
-    }
-
-    const { location: cityCenter } = props.currentCity;
-    const mapProperties = {
-      center: [
-        cityCenter.latitude,
-        cityCenter.longitude,
-      ] as LatLngTuple,
-      zoom: cityCenter.zoom,
-    };
-
-    const layer = new TileLayer(
-      'https://{s}.basemaps.cartocdn.com/rastertiles/voyager/{z}/{x}/{y}{r}.png',
-      {
-        attribution:
-          '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>'
-      }
-    );
-
-    const mapInstance = new LeafletGeoMap(nodeRef.current, mapProperties);
-    mapInstance.addLayer(layer);
-    setGeoMap(mapInstance);
-    isRenderedRef.current = true;
-  }, [props.locations, props.currentCity]);
+  const geoMap = useGeoMap(nodeRef, props.currentCity);
+  useGeoMapPins(geoMap, props.locations);
 
   return (
     <section
